Configuro pagos con el api Paypal PaypalController y cambio icono de Paypal

+++ website/database/migrations/2017_05_29_045911_cotizaciones.php
@@ -24,7 +24,7 @@ class Cotizaciones extends Migration
             $table->foreign('publicacion_id')
                 ->references('id')
                 ->on('publicaciones');
-            $table->integer('precio');
+            $table->decimal('precio', 10, 2);
             $table->integer('estado');
             $table->text('descripcion')->nullable();
             $table->text('ruta_entrega')->nullable();
